third_party/blink/web_tests/external/wpt/fetch/private-network-access/window-open-existing.tentative.https.window_include=from-public-expected.txt

This is a testharness.js-based test.
[FAIL] public to local: failed preflight.
  assert_equals: expected "timeout" but got "success"
[FAIL] public to local: missing CORS headers.
  assert_equals: expected "timeout" but got "no preflight received"
[FAIL] public to local: missing PNA header.
  assert_equals: expected "timeout" but got "no preflight received"
[FAIL] public to local: success.
  assert_equals: expected "success" but got "no preflight received"
[FAIL] public to private: failed preflight.
  assert_equals: expected "timeout" but got "success"
[FAIL] public to private: missing CORS headers.
  assert_equals: expected "timeout" but got "no preflight received"
[FAIL] public to private: missing PNA header.
  assert_equals: expected "timeout" but got "no preflight received"
[FAIL] public to private: success.
  assert_equals: expected "success" but got "no preflight received"
Harness: the test ran to completion.
